requirements.txt
streamlit  
streamlit-chat
google-generativeai